// types.ts
export enum LoadingStage {
    LOGIN = -1, // New initial stage
    LOGO = 0,
    LANDING = 2,
    HUB = 3,
    PROGRAM_DETAIL = 4, 
    ENROLLMENT = 5, 
    SHOP = 6, 
    LOW_TIMER_GAP = 7, 
    BLACK_BOX = 8, 
    LATEST_NEWS = 9, 
    DEVELOPER_EDITOR = 10, 
    OPERATING_HANDBOOK = 11, 
    TEAM_PAGE = 12,
    EXAMINATION_TERMINAL = 14
  }
  
  export enum PilotLevel {
    STUDENT = 1,
    PPL = 2,
    CPL = 3,
    MENTOR = 4
  }